add default api version to swagger
