+++ src/app/AlwaysOn.CatalogService/Startup.cs
@@ -1,4 +1,5 @@
 using AlwaysOn.CatalogService.Auth;
+using AlwaysOn.CatalogService.SwaggerHelpers;
 using AlwaysOn.Shared;
 using AlwaysOn.Shared.Interfaces;
 using AlwaysOn.Shared.Services;
@@ -61,6 +62,7 @@ namespace AlwaysOn.CatalogService
                 var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
                 c.IncludeXmlComments(xmlPath);
                 c.OperationFilter<ApiKeyFilter>(); // Custom parameter in Swagger for API Key-protected operations
+                c.OperationFilter<VersionParameterFilter>(); // Custom add default value for version parameter in Swagger
             });
 
             services.AddCors();
